Prevent default for every shortcut
This change is motivated by the fact that Firefox captures the "/", and sends opens the "find on page" feature. This is a great feature, but if the page has defined a shortcut for "/", the page's shortcut should take precedence. Along those lines, prevent default should be applied to every defined shortcut